fix: corrige bugs no schema, modelos, serviços e configuração do banco

+++ backend/src/main/java/com/libraryproject/model/ItemPedido.java
@@ -1,6 +1,18 @@
 package com.libraryproject.model;
 
-import jakarta.persistence.*;
+import java.math.BigDecimal;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "ItemPedido")
@@ -22,7 +34,7 @@ public class ItemPedido {
     private Integer quantidade = 1;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private double precoUnitario;
+    private BigDecimal precoUnitario;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 5)
@@ -60,11 +72,11 @@ public class ItemPedido {
         this.quantidade = quantidade;
     }
 
-    public double getPrecoUnitario() {
+    public BigDecimal getPrecoUnitario() {
         return precoUnitario;
     }
 
-    public void setPrecoUnitario(double precoUnitario) {
+    public void setPrecoUnitario(BigDecimal precoUnitario) {
         this.precoUnitario = precoUnitario;
     }
